:recycle:  keep functino parseGroupByOption into core/config
