+++ packages/utils/src/group.js
@@ -1,22 +1,3 @@
-const DEFAULT_GROUP = "Miscellaneous";
-const OPTION_REGEX =
-  /^@(?<directive>\w+)\((?<field>\w+)(?:\|=(?<fallback>\w+))?\)/;
-
-function parseGroupByOption(groupOptions) {
-  if (typeof groupOptions !== "string") {
-    return undefined;
-  }
-
-  const parsedOptions = OPTION_REGEX.exec(groupOptions);
-
-  if (typeof parsedOptions === "undefined" || parsedOptions == null) {
-    throw new Error(`Invalid "${groupOptions}"`);
-  }
-
-  const { directive, field, fallback = DEFAULT_GROUP } = parsedOptions.groups;
-  return { directive, field, fallback };
-}
-
 function getGroups(rootTypes, groupByDirective) {
   const groups = {};
 
@@ -63,4 +44,4 @@ function getGroupName(type, groupByDirective) {
   return groupByDirective.fallback;
 }
 
-module.exports = { getGroupName, getGroups, parseGroupByOption };
+module.exports = { getGroupName, getGroups };
